requirements.txt
flask==2.3.3
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
faiss-cpu==1.7.4
werkzeug==2.3.7
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.9
sentence-transformers==3.4.1
python-dotenv==1.0.1 